Refactor: JPA 엔티티 패키지 경로 변경 #10
domain 하위의 새 model 패키지로 클래스 이동
