+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/command/AuthMemberCommandAdapter.java
@@ -4,7 +4,7 @@ import com.nextdoor.nextdoor.common.Adapter;
 import com.nextdoor.nextdoor.domain.auth.port.AuthMemberCommandPort;
 import com.nextdoor.nextdoor.domain.auth.service.dto.MemberCommandDto;
 import com.nextdoor.nextdoor.domain.auth.service.dto.MemberQueryDto;
-import com.nextdoor.nextdoor.domain.member.domain.Member;
+import com.nextdoor.nextdoor.domain.member.domain.model.Member;
 import com.nextdoor.nextdoor.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
