+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/fintech/domain/Account.java
@@ -2,7 +2,7 @@ package com.nextdoor.nextdoor.domain.fintech.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.nextdoor.nextdoor.domain.member.domain.Member;
+import com.nextdoor.nextdoor.domain.member.domain.model.Member;
 import jakarta.persistence.*;
 import lombok.*;
 import java.time.LocalDateTime;
