+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/fintech/service/AccountService.java
@@ -9,13 +9,12 @@ import com.nextdoor.nextdoor.domain.fintech.dto.RegistAccountResponseDto;
 import com.nextdoor.nextdoor.domain.fintech.event.RemittanceCompletedEvent;
 import com.nextdoor.nextdoor.domain.fintech.port.MemberQueryPort;
 import com.nextdoor.nextdoor.domain.fintech.repository.AccountRepository;
-import com.nextdoor.nextdoor.domain.member.domain.Member;
+import com.nextdoor.nextdoor.domain.member.domain.model.Member;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.support.TransactionTemplate;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Schedulers;
