+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/domain/fintech/service/FintechUserService.java
@@ -2,14 +2,13 @@ package com.nextdoor.nextdoor.domain.fintech.service;
 
 import com.nextdoor.nextdoor.domain.auth.port.AuthFintechCommandPort;
 import com.nextdoor.nextdoor.domain.fintech.client.SsafyApiClient;
-import com.nextdoor.nextdoor.domain.member.domain.Member;
+import com.nextdoor.nextdoor.domain.member.domain.model.Member;
 import com.nextdoor.nextdoor.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
 import reactor.core.scheduler.Schedulers;
 
-import java.time.LocalDateTime;
 import java.util.Map;
 
 @Service
