+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/query/MemberQueryAdapter.java
@@ -2,8 +2,8 @@ package com.nextdoor.nextdoor.query;
 
 import com.nextdoor.nextdoor.common.Adapter;
 import com.nextdoor.nextdoor.domain.fintech.port.MemberQueryPort;
-import com.nextdoor.nextdoor.domain.member.domain.Member;
-import com.nextdoor.nextdoor.domain.member.domain.QMember;
+import com.nextdoor.nextdoor.domain.member.domain.model.Member;
+import com.nextdoor.nextdoor.domain.member.domain.model.QMember;
 import com.nextdoor.nextdoor.domain.member.repository.MemberRepository;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
