+++ backend/nextdoor/src/main/java/com/nextdoor/nextdoor/query/MemberUuidQueryAdapter.java
@@ -1,7 +1,7 @@
 package com.nextdoor.nextdoor.query;
 
 import com.nextdoor.nextdoor.common.Adapter;
-import com.nextdoor.nextdoor.domain.member.domain.QMember;
+import com.nextdoor.nextdoor.domain.member.domain.model.QMember;
 import com.nextdoor.nextdoor.domain.rentalreservation.domain.model.QRentalReservation;
 import com.nextdoor.nextdoor.domain.rentalreservation.domain.model.RentalReservation;
 import com.nextdoor.nextdoor.domain.rentalreservation.domain.exception.NoSuchRentalException;
